fix: widen UI UserData from uint32_t to uintptr_t to prevent pointer truncation on arm64
On arm64 macOS, pointers are 8 bytes but all UI control UserData fields
were uint32_t (32-bit). Storing a heap pointer via Set_Entry_Data() would
silently truncate the top 32 bits; retrieving it and casting back to a
pointer produced a garbage address, causing crashes (e.g. StringClass::
Free_String with address 0x00000000002809d0 when closing Movie Options).

Fix: widen UserData storage to uintptr_t throughout the UI widget layer
and update all callsites to remove the now-unnecessary (DWORD)(uintptr_t)
truncating double-casts.

Files changed:
- wwui/listctrl.h + .cpp: ListEntryClass::UserData, Get/Set_User_Data,
  ListCtrlClass::Get/Set_Entry_Data, Sort/Quick_Sort/Default_Sort_Callback
- wwui/mapctrl.h + .cpp: MapMarkerClass::UserData, Get/Set_Marker_Data
- wwui/dropdownctrl.h + .cpp: _ENTRY::user_data, Get/Set_Item_Data
- wwui/comboboxctrl.h: forwarding Get/Set_Item_Data wrappers
- 8 dialog headers: sort callback declarations (LISTCTRL_SORT_CALLBACK)
- 9 dialog cpps: sort callback implementations + truncating cast removal

+++ original/Code/Commando/dlgcncbattleinfo.h
@@ -44,7 +44,7 @@ private:
 	void		Populate_Player_List (ListCtrlClass *list_ctrl, int team_id);
 	void		Build_Player_Display_Name(const cPlayer* player, WideStringClass& outName);
 	
-	static int CALLBACK ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+	static int CALLBACK ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 	
 	///////////////////////////////////////////////////////////////////
 	//	Private member data
