+++ original/Code/Commando/dlgevaviewertab.h
@@ -133,7 +133,7 @@ protected:
 	////////////////////////////////////////////////////////////////
 	//	Static methods
 	////////////////////////////////////////////////////////////////
-	static int CALLBACK ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uint32_t user_param);
+	static int CALLBACK ListSortCallback (ListCtrlClass *list_ctrl, int item_index1, int item_index2, uintptr_t user_param);
 
 	////////////////////////////////////////////////////////////////
 	//	Protected member data
