+++ original/Code/wwui/dropdownctrl.h
@@ -43,8 +43,8 @@ public:
 	void					Delete_String (int index);
 	int					Find_String (const WCHAR* string);
 	int					Find_Closest_String(const WCHAR* string);
-	void					Set_Item_Data (int index, uint32_t data);
-	uint32_t				Get_Item_Data (int index);
+	void					Set_Item_Data (int index, uintptr_t data);
+	uintptr_t				Get_Item_Data (int index);
 	void					Reset_Content (void);
 	
 	//
@@ -101,7 +101,7 @@ protected:
 	typedef struct _ENTRY
 	{
 		WideStringClass	text;
-		uint32_t				user_data;
+		uintptr_t				user_data;
 
 		bool operator== (const _ENTRY &src)	{ return false; }
 		bool operator!= (const _ENTRY &src)	{ return true; }
@@ -109,7 +109,7 @@ protected:
 		_ENTRY (void)	:
 			user_data (0) {}
 
-		_ENTRY (const WCHAR *_text, uint32_t _data)	:
+		_ENTRY (const WCHAR *_text, uintptr_t _data)	:
 			text (_text), user_data (_data) {}
 
 	} ENTRY;
