+++ original/Code/wwui/mapctrl.h
@@ -44,13 +44,13 @@ public:
 	const Vector3 &	Get_Position (void) const				{ return Position; }
 	const WCHAR *		Get_Name (void) const					{ return Name; }
 	const RectClass &	Get_Rect (void) const					{ return UVRect; }
-	uint32_t				Get_User_Data (void)	const				{ return UserData; }
+	uintptr_t				Get_User_Data (void)	const				{ return UserData; }
 	uint32_t				Get_Color (void) const					{ return Color; }
 	
 	void					Set_Position (const Vector3 &pos)	{ Position = pos; }
 	void					Set_Name (const WCHAR *name)			{ Name = name; }
 	void					Set_Rect (const RectClass &rect)		{ UVRect = rect; }
-	void					Set_User_Data (uint32_t data)			{ UserData = data; }
+	void					Set_User_Data (uintptr_t data)			{ UserData = data; }
 	void					Set_Color (uint32_t color)				{ Color = color; }
 
 protected:
@@ -61,7 +61,7 @@ protected:
 	Vector3				Position;
 	WideStringClass	Name;
 	RectClass			UVRect;
-	uint32_t				UserData;
+	uintptr_t				UserData;
 	uint32_t				Color;
 };
 
@@ -119,8 +119,8 @@ public:
 	//	Marker control
 	//
 	int				Add_Marker (const WCHAR *name, const Vector3 &pos, const RectClass &uv_rect, int color = 0xFFFFFFFF);
-	uint32_t			Get_Marker_Data (int index);
-	void				Set_Marker_Data (int index, uint32_t user_data);
+	uintptr_t			Get_Marker_Data (int index);
+	void				Set_Marker_Data (int index, uintptr_t user_data);
 	void				Remove_Marker (int index);
 
 	//
